+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 3.0.0 - 2017-07-1
 
+- Change tests to Tape/Tap instead of Jest
 - Add Schema parameters -- allows the user to enter TMS, Quadkey or XYZ as tile format and returns that format as well.
 - **BREAKING CHANGE**: Default Tile Schema is XYZ (before it was TMS)
 - **BREAKING CHANGE**: `hashes()` output => `Promise<Set(number)>`